requirements.txt
beautifulsoup4
google-api-python-client
pandas
pytubefix
requests
python-dotenv
sentence-transformers
